requirement.txt
Django==5.0.6
djangorestframework==3.15.2
requests==2.32.3
google-auth==2.35.0
firebase==4.0.1
firebase-admin==6.5.0